# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for Auth Service.",
        "title": "Auth Service API3",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:4444",
    "paths": {
...
        "/user/createEcopoints": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ushbu endpoint foydalanuvchi uchun eko ballarini yaratish uchun ishlatiladi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Foydalanuvchi eko ballarini yaratish",
                "parameters": [
                    {
                        "description": "Foydalanuvchi eko ballari ma'lumotlari",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.CreateEcoPoints"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Yaratish muvaffaqiyatli",
                        "schema": {
                            "$ref": "#/definitions/users.InfoUserEcoPoints"
                        }
                    },
                    "400": {
                        "description": "Xato so'rov",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/user/deleteProfile/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ushbu endpoint foydalanuvchi profilini o'chirish uchun ishlatiladi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Foydalanuvchi profilini o'chirish",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Foydalanuvchi ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "O'chirish muvaffaqiyatli",
                        "schema": {
                            "$ref": "#/definitions/users.Status"
                        }
                    },
                    "400": {
                        "description": "Xato so'rov",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/user/getAllUsers": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ushbu endpoint barcha foydalanuvchilarni olish uchun ishlatiladi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Barcha foydalanuvchilarni olish",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Foydalanuvchi to'liq ismi",
                        "name": "full_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Foydalanuvchilar ro'yxati",
                        "schema": {
                            "$ref": "#/definitions/users.Users"
                        }
                    },
                    "400": {
                        "description": "Xato so'rov",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/user/getProfile/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ushbu endpoint asosida foydalanuvchi profilini olish uchun ishlatiladi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Foydalanuvchi profilini olish",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Foydalanuvchi identifikatori",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/users.UserInfo"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/user/historyUserPoint/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ushbu endpoint foydalanuvchi uchun eko ballari tarixini olish uchun ishlatiladi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Foydalanuvchi eko ballari tarixini olish",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarix ro'yxati",
                        "schema": {
                            "$ref": "#/definitions/users.Histories"
                        }
                    },
                    "400": {
                        "description": "Xato so'rov",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/user/updateProfile": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ushbu endpoint foydalanuvchi profilini yangilash uchun ishlatiladi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Foydalanuvchi profilini yangilash",
                "parameters": [
                    {
                        "description": "Profil yangilash so'rovi",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.ProfileUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Yangilash muvaffaqiyatli",
                        "schema": {
                            "$ref": "#/definitions/users.UpdateResponse"
                        }
                    },
                    "400": {
                        "description": "Xato so'rov",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/user/userEcopoints/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Ushbu endpoint foydalanuvchi eko ballarini olish uchun ishlatiladi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Foydalanuvchi eko ballarini olish",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Foydalanuvchi ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Foydalanuvchi eko ballari",
                        "schema": {
                            "$ref": "#/definitions/users.UserEcoPoints"
                        }
                    },
                    "400": {
                        "description": "Xato so'rov",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "users.CreateEcoPoints": {
            "type": "object",
            "properties": {
                "ecoPoints": {
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "users.Histories": {
            "type": "object",
            "properties": {
                "histories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/users.HistoryUserPoints"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "users.HistoryUserPoints": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "points": {
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                },
                "transactionId": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "users.InfoUserEcoPoints": {
            "type": "object",
            "properties": {
                "addedPoints": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "ecoPoints": {
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "users.ProfileUpdate": {
            "type": "object",
            "properties": {
                "bio": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "users.Status": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "users.UpdateResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "users.User": {
            "type": "object",
            "properties": {
                "ecoPoints": {
                    "type": "integer"
                },
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "users.UserEcoPoints": {
            "type": "object",
            "properties": {
                "ecoPoints": {
                    "type": "integer"
                },
                "lastUpdated": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "users.UserInfo": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "ecoPoints": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "users.Users": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/users.User"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}